backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
python-dotenv==1.0.0
aiofiles==23.2.1
openai==0.28.1
httpx==0.24.1
